+++ src/matrix/matrix.h
@@ -37,7 +37,7 @@ typedef struct s_shearing
 
 // m4default: Get the default matrix
 // @return The default matrix
-t_matrix4	m4default();
+t_matrix4	m4default(void);
 // m4mul: Multiply two matrices
 // @param m The first matrix
 // @param n The second matrix
