+++ src/vectors/vectors_utils.c
@@ -1,6 +1,6 @@
 #include "vectors.h"
 
-t_vector3   vreflect(t_vector3 v, t_vector3 normal)
+t_vector3	vreflect(t_vector3 v, t_vector3 normal)
 {
-    return (vsub(v, vmul(normal, 2 * vdot(v, normal))));
+	return (vsub(v, vmul(normal, 2 * vdot(v, normal))));
 }
